Refactor to improve readability
Use `await` in `extension.ts`.
Move some functionality out of `extension.ts`.
Move some TypeScript files into subfolders.
Fix `.editorconfig`.

Fixes #418
