+++ .editorconfig
@@ -1,9 +1,9 @@
 root = true
 
-[*.ts]
-[*.js]
-indent_style = space
-indent_size = 2
 charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = false
+
+[*.{ts,js,json,md}]
+indent_style = space
+indent_size = 2
