+++ src/server/java/javaServerStarter.ts
@@ -2,8 +2,8 @@ import * as os from 'os';
 import * as path from 'path';
 import { ExtensionContext, workspace } from 'vscode';
 import { Executable } from 'vscode-languageclient';
-import { RequirementsData } from './requirements';
-import { getJavaagentFlag, getKey, getXMLConfiguration, IS_WORKSPACE_VMARGS_XML_ALLOWED, xmlServerVmargs } from './settings';
+import { RequirementsData } from '../requirements';
+import { getJavaagentFlag, getKey, getXMLConfiguration, IS_WORKSPACE_VMARGS_XML_ALLOWED, xmlServerVmargs } from '../../settings/settings';
 const glob = require('glob');
 
 declare var v8debug;
