+++ src/settings/externalXmlSettings.ts
@@ -1,0 +1,24 @@
+import { XMLFileAssociation } from "../api/xmlExtensionApi";
+
+/**
+ * Represents vscode-xml settings that other vscode extensions can contribute to
+ */
+export class ExternalXmlSettings {
+
+  private _xmlCatalogs: string[]
+  private _xmlFileAssociations: XMLFileAssociation[]
+
+  constructor() {
+    this._xmlCatalogs = [];
+    this._xmlFileAssociations = [];
+  }
+
+  get xmlCatalogs(): string[] {
+    return this._xmlCatalogs;
+  }
+
+  get xmlFileAssociations(): any[] {
+    return this._xmlFileAssociations;
+  }
+
+}
